This is a big commit. I will need to start making smaller commits more often.
New Features:

- Router, replacing the Express router
- Localization facilities
- YAML resources

Fixes:

- Views context sometimes was the global. I recall it was nested views that had the problem. Fixed it by binding the view to the caller's context.

Introduced a router to be used in place of Express's router. This will likely lead to other changes, such as own request and response objects. The router was meant to introduce the ability to replace callbacks for routes, such as when we reload them when the code changes.